video_files filled and added to interface

+++ db/migrate/20140315211720_create_video_files.rb
@@ -7,7 +7,7 @@ class CreateVideoFiles < ActiveRecord::Migration
       t.text :external_url
       t.string :quality
       t.integer :year
-      t.integer :size
+      t.float :size
       t.belongs_to :content
 
       t.timestamps
